Split benchmark CLI into focused modules
