+++ crates/karva_benchmark/src/bin/karva_benchmark/metric.rs
@@ -1,0 +1,148 @@
+use clap::ValueEnum;
+use serde::{Deserialize, Serialize};
+
+pub const MATERIAL_CHANGE_PERCENT: f64 = 1.0;
+
+#[derive(Debug, Clone, Copy, Deserialize, Eq, PartialEq, Serialize, ValueEnum)]
+#[serde(rename_all = "kebab-case")]
+pub enum BenchmarkMetric {
+    WallTime,
+    Memory,
+}
+
+impl BenchmarkMetric {
+    pub fn marker(self) -> &'static str {
+        match self {
+            Self::WallTime => "<!-- karva-benchmark-comparison -->",
+            Self::Memory => "<!-- karva-memory-benchmark-comparison -->",
+        }
+    }
+
+    pub fn mode_label(self) -> &'static str {
+        match self {
+            Self::WallTime => "WallTime",
+            Self::Memory => "Memory",
+        }
+    }
+
+    pub fn warning_label(self) -> &'static str {
+        match self {
+            Self::WallTime => "wall-time",
+            Self::Memory => "peak-memory",
+        }
+    }
+
+    pub fn report_context(self) -> &'static str {
+        match self {
+            Self::WallTime => {
+                "Each benchmark compares median CLI wall time on one GitHub Actions runner, alternating install order. Runs warm the duration cache before measuring and include default per-test status output. Lower is better."
+            }
+            Self::Memory => {
+                "Each benchmark compares median peak RSS for the installed Karva CLI on one GitHub Actions runner, alternating install order. Runs warm the duration cache before measuring and are configured per project. Lower is better."
+            }
+        }
+    }
+
+    pub fn regression_verdict(self) -> &'static str {
+        match self {
+            Self::WallTime => "Merging this PR may alter performance",
+            Self::Memory => "Merging this PR may increase memory usage",
+        }
+    }
+
+    pub fn improvement_verdict(self) -> &'static str {
+        match self {
+            Self::WallTime => "Merging this PR improves performance",
+            Self::Memory => "Merging this PR reduces memory usage",
+        }
+    }
+
+    pub fn unchanged_verdict(self) -> &'static str {
+        match self {
+            Self::WallTime => "Merging this PR will not alter performance",
+            Self::Memory => "Merging this PR will not alter memory usage",
+        }
+    }
+
+    pub fn format_value(self, value: f64) -> String {
+        match self {
+            Self::WallTime => format_seconds(value),
+            Self::Memory => format_peak_rss_kib(value),
+        }
+    }
+}
+
+pub fn median(values: &[f64]) -> f64 {
+    let mut sorted = values.to_vec();
+    sorted.sort_by(f64::total_cmp);
+    let midpoint = sorted.len() / 2;
+
+    if sorted.len().is_multiple_of(2) {
+        f64::midpoint(sorted[midpoint - 1], sorted[midpoint])
+    } else {
+        sorted[midpoint]
+    }
+}
+
+pub fn percent_change(baseline: f64, candidate: f64) -> f64 {
+    ((candidate - baseline) / baseline) * 100.0
+}
+
+pub fn format_percent(percent: f64) -> String {
+    if percent.is_sign_positive() {
+        format!("+{percent:.1}%")
+    } else {
+        format!("{percent:.1}%")
+    }
+}
+
+pub fn trend(percent_change: f64) -> &'static str {
+    if percent_change <= -MATERIAL_CHANGE_PERCENT {
+        "faster"
+    } else if percent_change >= MATERIAL_CHANGE_PERCENT {
+        "slower"
+    } else {
+        "flat"
+    }
+}
+
+pub fn is_material_change(percent_change: f64) -> bool {
+    percent_change.abs() >= MATERIAL_CHANGE_PERCENT
+}
+
+pub fn trend_marker(percent_change: f64) -> &'static str {
+    match trend(percent_change) {
+        "faster" => ":zap:",
+        "slower" => ":x:",
+        _ => ":white_check_mark:",
+    }
+}
+
+fn format_seconds(seconds: f64) -> String {
+    if seconds < 1.0 {
+        format!("{:.1} ms", seconds * 1000.0)
+    } else {
+        format!("{seconds:.3} s")
+    }
+}
+
+fn format_peak_rss_kib(peak_rss_kib: f64) -> String {
+    if peak_rss_kib < 1024.0 {
+        format!("{peak_rss_kib:.0} KiB")
+    } else {
+        format!("{:.1} MiB", peak_rss_kib / 1024.0)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::trend;
+
+    #[test]
+    fn trend_uses_material_change_threshold() {
+        assert_eq!(trend(-1.0), "faster");
+        assert_eq!(trend(1.0), "slower");
+        assert_eq!(trend(0.9), "flat");
+        assert_eq!(trend(-0.9), "flat");
+    }
+}
